requirements.txt
livekit
livekit-agents
livekit-plugins-google
python-dotenv
asyncio
keyboard  # Optional: for real-time spacebar control